+++ backend/docker-compose.yaml
@@ -11,3 +11,7 @@ services:
       - "5432:5432"
     volumes:
       - ./postgres:/var/lib/postgresql/data
+  redis:
+    image: redis:4.0
+    ports:
+      - "6379:6379"
