+++ backend/officehoursqueue/__init__.py
@@ -1,0 +1,4 @@
+from officehoursqueue.celery import app as celery_app
+
+
+__all__ = ("celery_app",)
