+++ backend/officehoursqueue/celery.py
@@ -1,0 +1,19 @@
+import os
+
+from celery import Celery
+from django.conf import settings
+
+
+# set the default Django settings module for the 'celery' program.
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "officehoursqueue.settings.development")
+
+app = Celery("OHQ", broker=settings.MESSAGE_BROKER_URL)
+
+# Using a string here means the worker doesn't have to serialize
+# the configuration object to child processes.
+# - namespace='CELERY' means all celery-related configuration keys
+#   should have a `CELERY_` prefix.
+app.config_from_object("django.conf:settings", namespace="CELERY")
+
+# Load task modules from all registered Django app configs.
+app.autodiscover_tasks()
